Add files via upload
This file elaborates the use of Planetary Computer's landsat dataset which is on microsoft cloud storage. Here the landsat-c2-l2 (level 2)  COG data is used, First the environment is set by importing all the required libraries. Then the data is accessed by using the url, the search for the aoi is done, the RGB image along with it's metadata is found. Again some of the indexes are also found.